% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISASP_Package.R
\docType{package}
\name{ISASP-package}
\alias{ISASP-package}
\title{Easy Reporting of Iowa Statewide Assessment of Student Progress}
\description{
The Iowa Statewide Assessment of Student Progress (ISASP) has replaced the Iowa Assessment as the standard test for students in Iowa
The reports generated by the previous Assessment and not all directly available at this time.
This package aims to make it easier to generate some of these reports by providing helpful functions to aid in the cleaning and extraction of key values.
}
\author{
Joshua De la Burere (joshua@delabj.com)
}
