% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proficency_percents.R
\name{proficency_pct}
\alias{proficency_pct}
\title{Calculate Proficiency Percentages}
\usage{
proficency_pct(df, col = ELAAchLvl, by = NULL, levels = c("A", "P", "N"))
}
\arguments{
\item{df}{A data frame from the ISASP report}

\item{col}{the column name to calculate proficiency percentages from}

\item{by}{the column to aggregate by (School, Grade, Etc)}

\item{A}{list of levels as they appear in the data. This should not typically have to change.}
}
\value{
a data frame with aggregated percentages by the chosen grouping
}
\description{
easily calculate proficiency levels for ISASP
}
