% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_ISASP_data.R
\name{student_subtest_score}
\alias{student_subtest_score}
\title{Extract Sub-Test Scores By Student}
\usage{
student_subtest_score(
  df,
  label_col = NULL,
  pct_col = NULL,
  points_col = NULL,
  raw_score_col = NULL,
  student_id = StateID,
  grade = Grade
)
}
\arguments{
\item{df}{A data frame}

\item{label_col}{Column name of the test label for the sub-test}

\item{pct_col}{Column name of the test percent for the sub-test}

\item{points_col}{Column name of the total points possible column for the sub-test}

\item{raw_score_col}{Column name of the raw Score for the sub-test}
}
\value{
A 6 column data frame containing the columns \code{student_id}, \code{grade}, \code{test_label}, \code{raw_score}, \code{pct_correct}, and \code{points_possible}
}
\description{
Extract a particular column of sub-test scores.
Mostly useful for stacking said sub-test scores and rejoining with student data later on.
}
